requirements.txt
faker>=24.0.0
pydantic>=2.7.0
pyyaml>=6.0.1
reportlab>=4.0.9
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.26.0
boto3>=1.34.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
